Created C# & VB.NET Examples for new Release 4.0.0 Release
