+++ Examples/VisualBasic/packages.config
@@ -1,4 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Aspose.Imaging" version="3.7.0" targetFramework="net40" />
+  <package id="Aspose.Imaging" version="4.0.0" targetFramework="net40" />
 </packages>
